Added converter as a service + working in loops
